Second read through

+++ src/test.rs
@@ -1,4 +1,4 @@
-pub fn sqrt_n_over_n(n: u32) -> f64 {
-    let f = n as f64;
+pub fn sqrt_n_over_n(n: u32) -> f32 {
+    let f = n as f32;
     f.sqrt() / f
 }
